Gas cost functions return uint32
A change was made to the complete repo to alter data types of gas to uint32. This
caused tests in this branch to fail as they were expecting uint64.

Also applied some changes to integration tests that were failing when using the
-short flag.

+++ core/services/keeper/common.go
@@ -32,6 +32,6 @@ type Config interface {
 }
 
 type RegistryGasChecker interface {
-	KeeperRegistryCheckGasOverhead() uint64
-	KeeperRegistryPerformGasOverhead() uint64
+	KeeperRegistryCheckGasOverhead() uint32
+	KeeperRegistryPerformGasOverhead() uint32
 }
